freelancefinder/requirements/base.txt
bleach==2.0.0
Celery[redis]==4.0.2
Django==1.10.6
django-allauth==0.31.0
django-braces==1.11.0
django-crispy-forms==1.6.1
django-environ==0.4.1
django-log-request-id==1.3.2
django-model-utils==2.6.1
django-taggit==0.22.0
django-timezone-utils==0.11
django_celery_beat==1.0.1
django_logtail==0.3.0
feedparser==5.2.1
future==0.16.0
gunicorn==19.7.1
haxor==0.3.1
markdown==2.6.8
maya==0.1.8
nltk==3.2.2
praw==4.4.0
psycopg2==2.7.1
python-json-logger==0.1.7
pytzdata==2017.1
